refactor: AuthProvider component
